Fix corner radius type mismatch in CivitaiProgressButton
Convert cornerRadius from Int to Float in drawRoundRect call to fix build error.

+++ app/src/main/java/com/example/kennys_dokidoki_wallpaper/CivitaiProgressButton.kt
@@ -31,7 +31,7 @@ class CivitaiProgressButton @JvmOverloads constructor(
         if (fraction != null && fraction > 0f) {
             canvas.drawRoundRect(
                 0f, 0f, width * fraction, height.toFloat(),
-                cornerRadius, cornerRadius, fillPaint
+                cornerRadius.toFloat(), cornerRadius.toFloat(), fillPaint
             )
         }
     }
